+++ src/lib/platform/OSXMediaKeySupport.mm
@@ -67,6 +67,12 @@ convertNXKeyTypeToKeyID(uint32_t const type)
 	case NX_KEYTYPE_SOUND_DOWN:
 		id = kKeyAudioDown;
 		break;
+	case NX_KEYTYPE_BRIGHTNESS_UP:
+		id = kKeyBrightnessUp;
+		break;
+	case NX_KEYTYPE_BRIGHTNESS_DOWN:
+		id = kKeyBrightnessDown;
+		break;
 	case NX_KEYTYPE_MUTE:
 		id = kKeyAudioMute;
 		break;