+++ src/lib/platform/OSXScreen.mm
@@ -2456,7 +2456,8 @@ OSXScreen::onKey(CGEventRef event)
 		// get old and new modifier state
 		KeyModifierMask oldMask = getActiveModifiers();
 		KeyModifierMask newMask = m_keyState->mapModifiersFromOSX(macMask);
-		m_keyState->handleModifierKeys(getEventTarget(), oldMask, newMask);
+		m_keyState->handleModifierKeysEx(getEventTarget(), oldMask, newMask,
+										 virtualKey, macMask);
 
 		// if the current set of modifiers exactly matches a modifiers-only
 		// hot key then generate a hot key down event.